Collisions, Lives, Assets, Menu upgrades
Collisions now occur between the player and the enemy.
This is shown to the player through the hearts on the screen, representing their remaining lives.
The player's score is also shown on the screen. The score is dependent on how long the player survives, not how many obstacles they dodge.

The player is now represented as a clownfish, and the enemy as a shark. The clownfish is animated, and the shark's animations will be added in future.

The main menu now has a logo and a settings button, which opens the settings menu. The settings menu still needs work, but it will include a toggle for the type of controls for movement (gyroscope or touch).