Movement Type Toggle, UI Navigation
Gyroscopic movement has been implemented, so movement type can be toggled between touch and gyroscope.

The UI can be navigated fully, and out of all features shown, only colour blind settings have not been implemented.

Data persistence is present in the form of saving the player's movement control settings, although there is a visual glitch when first opening the game where gyroscope is shown to be toggled on when it is not active.