FMOD installed, Menu Code Adjusted
FMOD has been installed but is not yet being used.

MenuSceneController had a lot of code related to the main menu, but some code was needed for pause menu. Most of the code has been moved into the MainMenu class, so that the MenuSceneController can be used in multiple scenes.

PauseMenu class holds code for when the game is paused